requirement.txt
asgiref==3.5.2
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.6
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
